import api: fix duplication of nested patient resources
Sending the same patient resource more than once (either on accident or
for an attribute update) leads to the addresses, phone numbers and
business identifiers being duplicated, because we are generating a new
ID for them every time. We now ensure that these IDs are generated
deterministically so that the resource merge merely updates these values
rather than creating new instances.