Initialize ET dynamic forms 1 month in advance (#5289)
**Story card:**
[sc-11191](https://app.shortcut.com/simpledotorg/story/11191/initialize-all-dynamic-form-months-for-et)

## Because

All the dynamic forms are initialised on first of every month. In
Ethiopia, due to the difference in calendar, these forms appear around
20th of a month.

## This addresses

Disclaimer: We are going for a hack to fix this because accommodating
Ethiopian calendar in Simple will take up more than 2-3 months of work.

- We have decided to initialise the `current month` form in Ethiopia
instead of `previous month` form from next month onwards. That is, On
1st Nov, we will be initialising the form for `Nov-2023` in Ethiopia and
`Oct-2023` in other countries.
- We will have to manually initialise `Oct-2023` form for Ethiopia to
catch up with other countries.
---
- We have moved the initialisation code to a rake task
- We had to pass in a different `date` argument to initialise forms in
different countries.
  - With `runner` taking a `string` was not practical in our case
## Test instructions

- Run the rake task locally and verify if the forms are being created
    - `bundle exec rake questionnaires:initiliaze`

---------

Co-authored-by: akshat <olttwa@users.noreply.github.com>