import api: validate incorrect observation codes
Our OpenAPI schema validator is not flexible enough to capture some edge
cases for observation resources–namely, it is possible to pass in a
blood pressure observation, with the systolic or diastolic code
specified twice, which will cause a failure at the database constraint
level. We instead validate this at the time of a request so that the
user is immediately informed of this error.