import api: fix patient name extraction
There was a goof-up with name extraction. Staring at the deeply nested
semi-uniform patterns of the FHIR spec has caused semantic satiation in
this author which has introduced this bug in the patient importer, where
instead of trying to extract "text" for the name, the code is trying to
extract "value", which always returns `nil`.

This is promptly fixed, with some unit tests that should have been there
to catch this slip in the first place.