import api: validate incorrect observation codes (#5303)
# Why

Our OpenAPI schema validator is not flexible enough to capture some edge
cases for observation resources–namely, it is possible to pass in a
blood pressure observation, with the systolic or diastolic code
specified twice. This will cause a failure at the database constraint
level. We instead validate this at the time of a request so that the
user is immediately informed of this error, and the request is not
propagated any further.

# The fix

We add new validations in our Imports API validator class. Any
observation resource must satisfy either one of these conditions:
* It contains a single component that is a valid Blood Sugar measurement
LOINC code.
* It contains two components, one of which is a valid systolic BP
measurement and the other is a valid diastolic BP measurement.