import API: validate existence of patients on dependent resources
We previously allowed dangling references to patients from its
associated resources like appointments to allow integrators of the API
to send resources in any order they please. But experience with testing
the API on a sandbox has made it clear that the problem of dangling
patient references is more pressing than we initially assumed. This
change ensures that any dependent resource references a patient that
already exists in the Simple database. It returns an error similar to
the one about unmapped facility if this patient is not found.