import api: add data migration for incorrect translations
Since the business identifiers were created correctly, but the resources
associated with those patients were not, we add a data migration to
rectify the situation.

It turns out that we have been creating these dependent resources with
the `patient_id` set to the `id` of the business identifier, instead of
the `patient_id` of the business identifier. This migration swaps out
the `patient_id` of the dependent resource.