Import API: Fix incorrect patient IDs in patient-associated resources (#5320)
The patient identifiers were created incorrectly in the dependent
resources. We accidentally added a namespace prefix that's used to
identify the patient business identifier, rather than the patient itself
(which does not use any kind of prefix).

While refactoring the `translate_patient_id` method, it became clear
that this utility had no usage across multiple FHIR resource importers,
so we removed it and made the mixin interface smaller.

Our specs also need rectification, they used the wrong patient creation
logic.

Since the business identifiers were created correctly, but the resources
associated with those patients were not, we add a data migration to
rectify the situation.

It turns out that we have been creating these dependent resources with
the `patient_id` set to the `id` of the business identifier, instead of
the `patient_id` of the business identifier. This migration swaps out
the `patient_id` of the dependent resource.