Hotfix: Handle exception and log debug info while recording a visit on a treatment group membership (#5330)
**Story card:**
[sc-11504](https://app.shortcut.com/simpledotorg/story/11504)

This is a hotfix for the bug where, when recording a visit on a treatment
group membership, a NoMethodError is raised when the facility
association on a visit doesn't exist for whatever reason. Details of the preliminary investigation are captured
in the card.

This fix handles the exception and logs some debugging information.