import user: separate users from multiple orgs
The Import User was previously created by finding the first facility in
the database and assigning the user to that facility. This is a problem
in any instance that has more than one organization, since it will
affiliate the import user with whichever organization that first
facility belonged to.

We resolve this by requiring an organization ID when finding or creating
an Import User.