Fix XSS vulnerability on new district page (#5339)
**Story card:**
[sc-11561](https://app.shortcut.com/simpledotorg/story/11561/fix-xss-vulnerability-on-new-district-page)

## Because

The `Add block` input field on the `new district` page has an XSS
vulnerability due to the use of the
[.html](https://api.jquery.com/html/) JQuery method to insert user input
into HTML elements. So if the user input contains malicious code, it
will get executed.

A couple of other places with this vulnerability:
- The region search bar, used in a couple of places, like /my_facilities
- On the edit facility page (/admin/facilities > facility name > Edit
facility), in the search bar under the title `Medical officers available
for teleconsultation` when you check `Teleconsultation enabled?`

## This addresses

This PR changes all the [.html](https://api.jquery.com/html/) JQuery
methods to .text while inserting user input.
For hygiene, it also sanitizes input given to methods such as .prepend, .attr
and .append, that also injest html strings.