Evict patient from experiment if their visit facility is deleted (#5336)
If a facility on a visit is deleted, an error occurs while recording a
visit on a patient enrolled in an experiment (also called a 
membership). This error can also occurs in another part of the 
experiments code, when fetching contents to send a message.
These are handled in this commit in order to not disrupt the 
daily experiment runner.