Add new indexes on reporting_patient_states (#5349)
**Story card:**
[sc-11600](https://app.shortcut.com/simpledotorg/story/11600/optimize-queries-used-by-dhis2-exporter)

The current queries for our monthly DHIS2 exports are slow. We want to
add a couple of indexes to the reporting_patient_states table to improve
this.

This PR adds a migration to create four composite indexes:
- on month_date and assigned_facility_id
- on month_date and assigned_facility_region_id
- on month_date and registration_facility_id
- on month_date and registration_facility_region_id

We're accounting for both the facility and the facility region because
both these fields are used throughout the codebase and we don't want one
query to pick up the correct index and not another.