fix validation errors for import user numbers
Switching to an import user per organization introduced a bug where
creating a new Import User causes a bug because the phone number fails
the uniqueness validations. We no longer have a globally constant phone
number—we instead generate a random string as our fake phone number that
is unique to each organization and guaranteed to not collide again.