Automate AlphaSms balance check (#5365)
**Story card:**
[sc-11640](https://app.shortcut.com/simpledotorg/story/11640/automate-alphasms-balance-check)

We want to automate balance check in Bangladesh and alert on low
balance.

The reason the error message is being logged and not raised as an
exception is so we can add a Datadog alert on the log. The corresponding
BSNL task relies on Sentry to alert on an uncaught exception for low
balance, but we have to manually resolve the alert each time for it to
show up again next time. This is not an issue on Datadog.