Send balance and expiry metrics in the alphasms balance check rake task (#5379)
**Story card:**
[sc-11640](https://app.shortcut.com/simpledotorg/story/11640/automate-alphasms-balance-check)

## Because

- We want to continue to track this via Prometheus, and metrics are a
better way to do this than by logs
- We will be able to graph our usage with metrics

## Considerations

- Since we're still sending metrics via statsd and are planning to use
an exporter to translate these metrics to the Prometheus format, the
metrics in this PR are also being sent via Statsd. The format is
standard for both Statsd and Prometheus
- We could send the expiry days as a label on the balance metric instead
of as a second metric. The statsd exporter supports conversion of Statsd
tags into Prometheus labels. Keeping them as separate metrics for now
since the Prometheus migration's implementation details are unknown
- I've removed the older logs, will set up metric-based alerts on
Datadog until the move to Prometheus is complete