import api: deprecate expectedSupplyDuration from MedicationRequest
This field is not used internally for any of Simple's reporting; it is
an artifact of the now-unused teleconsultations feature. Accepting this
field only complicates our reporting because partners would have to now
be careful to not update the duration for existing medications, since
this would break reports that rely on the fact that medications are only
every mutated for the purpose of setting them inactive.

Asking partners to not update this field is not easily enforceable.
Therefore, we remove this field.