add OpenSRP exporter script for Gonaduwa migration
This is a one-off script that is meant to generate a JSON data dump for
one of our facilities in Sri Lanka to help the OpenSRP team conduct user
testing.

The code is intentionally rough around the edges, and very specific to
the export for the user test. We may write a more well-crafted and
well-tested script in the future if making OpenSRP-compatible FHIR
exports is something we might need outside of this one-off test. But for
now the focus is on speed, and the ability to rapidly incorporate
any changes needed for the OpenSRP export, since the new app doesn't yet
have a stable contract.