imports api: disallow non-array resources (#5391)
There was a bug where we allowed non-array values within the top-level
"resources" key of the Import API payload. This caused an exception to
be thrown later in the request lifecycle. This bug fix ensures any
non-array values fail at the controller layer and return the appropriate
error code (ie, 400 Bad Request).

**Story card:** [sc-12003]