import api: deprecate expectedSupplyDuration from MedicationRequest (#5390)
This field is not used internally for any of Simple's reporting; it is
an artefact of the now-unused teleconsultations feature. Accepting this
field only complicates our reporting because partners would have to now
be careful to not update the duration for existing medications since
this would break reports that rely on the assumption that medications
are only ever mutated to mark them as inactive.

Asking partners to not update this field ever is hard to enforce.
Therefore, we remove this field. To avoid breaking any existing
integrations, we simply drop the field while converting it from FHIR to
Simple's format.

This is technically a breaking change, but since we only have one
partner in production, we will inform them directly so they can make the
necessary changes on their backend.

I considered returning a warning informing the client about this
deprecation, but decided against it since it only burdens the controller
and validations with more code and given that we have one partner, it
didn't seem worth it to include it only for them.

This is how the deprecation looks like in the API docs:

![image](https://github.com/simpledotorg/simple-server/assets/24277692/a4a76a02-dd2d-4976-896c-01151b0c5387)


**Story card:** [sc-12022]