data migration: remove duplicate medical histories for ICDDRB
There were some duplicate medical histories that were imported even
shortly after we ran our data migration. We recreate another data
migration that does the same thing again, but deduplicates for all the
new facilities that received all this bad data.

If this migration looks familiar because it's 98% similar to the
previous dedupe migration. We duplicate this dedupe migration so that we
have a clear record of us having run something similar twice. Also the
facilities are different this time.