data migration: remove old invalid BUHS drugs
BUHS' integration with the import API had a bug which caused duplicate
drug data to be sent. We delete all the data sent before March 8 (which
is around the time the error was fixed) as per their request.
Thankfully, we don't have to be too exact with the date boundary, as
there was no drug data sent between 17th January to 11th March.

Since BUHS will resend all the drug data again, these should be safe to
delete.