add reassignment eligibility indicator to patient
This is a part of the CC reassignment feature for use in Bangladesh. We
add a new field on the patient model which helps address sync issues
from the Android side. The field is non-nullable with a default value of
"unknown". This value is not manipulated by the server by any means, it
only exists as a placeholder for the value evaluated on the Android
side.

This patch itself includes a DB migration to add this field, along with
relevant validations and backward-compatible patient sync model changes.