add reassignment eligibility to patient (#5423)
**Story card:** [sc-12577]

## Because

[Link to
PRD](https://docs.google.com/document/d/1R6NnMMYIqs0VfFlyT_LQdCoa3Vdcl3URS9tNn6X0EsI/edit?usp=sharing)

This is a part of the CC reassignment feature for use in Bangladesh. We
add a new field on the patient model which helps address sync issues
from the Android side. The field is non-nullable with a default value of
"unknown". This value is not manipulated by the server by any means, it
only exists as a placeholder for the value evaluated on the Android
side.

Rationale for having this on the server side and as an enum:

https://docs.google.com/document/d/1R6NnMMYIqs0VfFlyT_LQdCoa3Vdcl3URS9tNn6X0EsI/edit#heading=h.saxgwg5kcdd0

## This addresses

This patch includes a DB migration to add this field, relevant
validations, and backwards-compatible changes to the patient sync model.

## Test instructions

(I have already done this, so you don't have to, but in case you want
to, read on)

* Deploy this PR on QA (already done).
* Run the latest dev build of the Android App (and enable the feature
flag called `patient_reassignment_v0`).
* Select Bangladesh > UHC Krishna.
* The only patient in this UHC, Ramesh Babu, has met the eligibility
criteria (see PRD).
* Press back and select not now, and sync the changes.
* SSH in the QA server, the value of `eligible_for_reassignment` should
be set to "yes".
* Repeat this by changing any other criteria of eligibility (either by
reassigning or changing medicines or adding uncontrolled BP) and sync to
server.
* The value on server should now say "no".
* On syncing, the value should stay consistent, ie, at no point should
the flag get reset as the server is not touching this field's value.