spec: fix time-related flakiness
We have a bunch of specs in our suite that fail when run at odd hours
due to inconsistent application of timezone reporting helpers. One
common anti-pattern is to setup the test in UTC but make the assertions
within a block that sets the timezone to the configured reporting time
zone. We move the setup within this block where necessary.

In the spec for RegionSummarySchema, these helpers that aligned the spec
timezone with the database timezone were not used at all, we fix these
as well.