fix incorrect exporter mapping in Bangladesh
The DHIS2 disaggregated diabetes exporter relies on the
`bangladesh-production.yml` file to determine which data elements the
values should be sent to. The lookup would break in production because
the keys on the backend do not exactly match the keys in the YAML file.
We unify the names and make them consistent with the key names
convention that is in the hypertension exporter.