facility reports: speed up BP logs
The query that fetches the recent BP measurements on the facility
reports page does so with superfluous joins without using the facility
ID field on Blood Pressures which is indexed. We put this behind a
feature flag so we can monitor the difference in performance in
production.

We don't need any of the encounter or observation data, so we change the
query to not require any joins. This should help speed up our facility
report load times.

While we are at it, we also remove some other feature flags we had added
to debug the cause of slow loads.