Fix slow page loads for facility regions (#5444)
**Story card:**
[sc-12796](https://app.shortcut.com/simpledotorg/story/12796/investigate-slow-loading-facility-dashboard-bangladesh)

## Because

Our facility-level reports in Bangladesh have become very slow.

## This addresses

The query that fetches the recent BP measurements on the facility
reports page does so with superfluous joins without using the facility
ID field on Blood Pressures which is indexed. We put this behind a
feature flag so we can monitor the difference in performance in
production.

We don't need any of the encounter or observation data, so we change the
query to not require any joins. This should help speed up our facility
report load times.

While we are at it, we also remove some other feature flags we had added
to debug the cause of slow loads.

### Analyzed Query Plan: Before


![image](https://github.com/user-attachments/assets/9966413d-75bf-4ab7-8f60-390c11fc9b4d)

### Analyzed Query Plan: After


![image](https://github.com/user-attachments/assets/74939a9d-62a4-4619-8c35-2c2b7e649ebf)

## Test instructions

Enable the feature flag `fast_bp_logs` on flipper