Add index to patient follow-ups
Our facility reports are slow due to there being only a suboptimal
compound index on the follow ups matview. We add an index on the
facility ID which when testing on Sandbox offered better query plans and
performance characteristics—it used a Bitmap Index Scan along with a
Bitmap Heap Scan, instead of a slow Index Scan that read too many
blocks.