Expose request path in prometheus metrics (#5456)
**Story card:** [sc-XXXX](URL)

## Because

We want to be able to see the exact path (not just the action name and
controller for our slowest requests).

That way, we can tell whether a slow load is, for example, happening
within a district, organisation or facility-level region. Or even a
specific facility.

## This addresses

* Adds a custom label that takes the path information from the Rack
environment. Ref:
https://github.com/discourse/prometheus_exporter?tab=readme-ov-file#metrics-collected-by-rails-integration-middleware

## Results

Along with the changes to the Grafana dashboard to include the path
label, we get this (left-middle and middle tables):


![image](https://github.com/user-attachments/assets/9fe07ec9-f0a7-4457-8331-45de04bf2b0e)