add OpenSRP exporter script for Gonaduwa migration (#5474)
This is a one-off script that is meant to generate a JSON data dump for
one of our facilities in Sri Lanka to help the OpenSRP team conduct user
testing.

The code is intentionally rough around the edges, and very specific to
the export for the user test. We may write a more well-crafted and
well-tested script in the future if making OpenSRP-compatible FHIR
exports is something we might need outside of this one-off test. But for
now the focus is on getting this data out ASAP, along with the
flexibility to rapidly incorporate any changes needed for the OpenSRP
export, since the new app doesn't yet have a stable contract.

**Story card:**
[sc-13106](https://app.shortcut.com/simpledotorg/story/13106/share-a-dump-with-gonaduwa-patients)

## This addresses

* Adds a few services for exporting FHIR resources (in the format
OpenSRP understands, which was agreed upon based on a many months of
manual QA)