Fixing export bug in dhis2 export for ethiopia (#5489)
**Story card:**
[sc-14094](https://app.shortcut.com/simpledotorg/story/14094/fix-dhis2-export-job-bug-ethiopia)

## Because

DHIS2 ruby client that we use works for old version of DHIS2 (2.28). The
newer versions of DHIS2 has there API response structure changed.
Because of this it was raising error, even though import was successful,
and sidekiq job was going to retry queue.

## This addresses

Changing the export method for ethiopia to handle the response
correctly, and not raise error when the response is 200. Since
Bangladesh still is using older version of DHIS2, this change will only
affect Ethiopia.

## Test instructions

Run `rake "dhis2:ethiopia_export"` and monitor the response. After
successful DHIS2 import, it should not raise an error.

---------

Co-authored-by: Ayushi Agrawal <ayushiagrawal@RTSL-P172G93770.local>