Actually allow disabling Datadog (#5493)
It is impossible to disable Datadog because we don't coerce the
environment variable that toggles Datadog to boolean. This change
coerces the value into a Boolean.

This bug has led to a lot of log noise in environments where Datadog is
disabled because of the server attempting to connect to a Datadog
instances that do not exist.

"false" is true. And `nil` is short-circuited by the OR operator.