Fix: Daily experiments task failing due to method missing from MobitelDeliveryDetails (#5509)
**Story card:**
[sc-13928](https://app.shortcut.com/simpledotorg/story/13928/investigate-sms-issue-in-sri-lanka)

## Because

The monitoring stage in the experimentation daily rake task is failing
to send experiment notifications. Hence, the actual SMS sending part,
which happens after monitoring, was not triggered.
The error was from a missing method "result" in MobitelDeliveryDetail
class.

## This addresses

We are adding the "result" method, which is used to determine the status
of the notification. Since Mobitel does not give back the actual state
of the notification, this is being hardcoded to "sent," assuming the API
call to Mobitel is successful and the message was sent.

## Test instructions

If we are using Mobitel as an SMS vendor, then `rake
"experiments:conduct_daily" should complete without errors, and we
should see SMSes being sent out.

---------

Co-authored-by: Ayushi Agrawal <ayushiagrawal@RTSL-P172G93770.local>