refactor(OpenSRP Exporter): External Config (#5533)
**Story card:**
[sc-14814](https://app.shortcut.com/simpledotorg/story/14814/opensrp-exports-export-a-specific-time-window)

## Because

Whenever we want to do an export, we have to update the rake file in the
running instance with the information per export. This is technically
configuration within the script, and doesn't scale as much

## This addresses

Enabling people bring their external configuration for the export. This
change does not alter the functionality in the exports.

## Test instructions

- suite tests