Whitelabeling text and images in swagger api docs (#5597)
**Story card:**
[sc-15243](https://app.shortcut.com/simpledotorg/story/15243/the-name-of-the-application-simple-should-come-from-the-configuration)

## Because

Swagger API docs have mentions of "Simple" with the default Simple logo.
We need the ability to whitelabel them.

## This addresses

We are using rswag which generates the api docs based on the generated
swagger.json file. In this we are overriding the texts to a custom
controller and then rendering the json.
With this approach, if someone modifies the swagger.json file for adding
new resource, the controller will whitelabel the required descriptions.

## Test instructions

Set a value for FAVICON_URL, ENG_EMAIL_ID, APPLICATION_BRAND_NAME in the
env variables.
Restart the server and access "api-docs/index.html".
The docs should reflect the brand changes.

---------

Co-authored-by: Ayushi Agrawal <ayushiagrawal@RTSL-P172G93770.local>