build: Setup release build process (#5605)
**Story card:**
[sc-15268](https://app.shortcut.com/simpledotorg/story/15268/versioning-adopt-semantic-versioning)

## Because

We want to release intentionally, following our versioning discussions.

This is part of the work we're undertaking to ensure Simple is a product
people can install themselves. Our releases have been bleeding edge for
the past seven years thereabouts. With releases, we are able to deliver
deterministically to our users.

## This addresses

Updating Semaphore instructions on the agreed way we want to deploy
releases

## Test instructions

n/a