fix(Dr. Rai Report): Remove Quartierles from Progress Tab (#5662)
**Story card:**
[sc-16476](https://app.shortcut.com/simpledotorg/story/16476/high-not-showing-in-the-progress-tab-for-the-facility-when-action-plans-are-created-on-bd-production)

## Because

The progress tab does not load in BD-PROD. In a bid not to debug in
prod, the assumption behind this change is that there is much data to
crunch in BD. So things like `quarterlies` which aren't really needed
but are still computed may be adding to load time.

## This addresses

Removing quarterlies from the lite version of Dr. Rai Reports

## Test instructions

n.a