Moving cache warmer job to a rake task (#5673)
**Story card:**
[sc-16582](https://app.shortcut.com/simpledotorg/story/16582/move-regioncachewarmer-job-to-a-rake-task)

## Because

We are currently running Reports::RegionCacheWarmer.call through runner
which fails in production env because of Spring contexts not properly
loaded into it.
As a result the indicator percentage caches carry stale data and causes
discrepancy in the dashboard.

## This addresses

Moving the execution to a rake task

## Test instructions

bundle exec rake db:warm_up_cache 
This should run without errors and must trigger the sidekiq jobs for
clearing up the caches

Co-authored-by: Ayushi Agrawal <ayushiagrawal@RTSL-P172G93770.local>