feat(Dr. Rai Reports): Future Plans (#5675)
**Story card:**
[sc-16603](https://app.shortcut.com/simpledotorg/story/16603/make-future-plans)

## Because

People may want to plan ahead

## This addresses

Allowing the next quarter to be visible in the component, but only
interacted with a month before the quarter finishes.

This also resulted some clean-up of the code — removing the extra region
summaries used to get the periods from the code path — hopefully
resulting in faster load times. This was replaced with the new
`Period.quarters_between` which does what it's called.

## Test instructions

suite tests