adding logic for doing partitioned refresh weekly (#5701)
**Story card:**
[sc-16751](https://app.shortcut.com/simpledotorg/story/16751/update-refresh-frequeny-for-weekly-report-refresh)

## Because

India is set to do the reports refresh once per week. Our refresh for
partitioned table assumes that it runs daily and hence just refreshes
for two months (current month and one past month)
This would mean that in India we will miss refreshing data for some
older months.

## This addresses

Adding a new environment variable REPORTING_REFRESH_FREQUENCY which when
set to "weekly" will then refresh for all the months, corresponding to
past 7 days, which would ideally have been refreshed if it was running
daily.
For e.g. Let's assume it is running on 24 Sep 2025, then it assumes then
it was last run on 17 Sep 2025, and now we have to a refresh for months
that correspond to 17, 18, 19, 20, 21, 22, 23 and 24 Sep 2025.
In this case months refreshed would be Sep 2025, Aug 2025, Aug 2024, Sep
2024, Oct 2024 and Nov 2024

## Test instructions
Set REPORTING_REFRESH_FREQUENCY="weekly"
run rake "db:refresh_reporting_views". 
It would then refresh for all the months corresponding to the last week.
This can be verified in entries under
"simple_reporting.simple_reporting_runs" table

Enter detailed instructions for how to test this PR...

---------

Co-authored-by: Ayushi Agrawal <ayushiagrawal@RTSL-P172G93770.local>